Add warning to avoid accidental overwrite

+++ deploy.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 DEPLOY_FILES="assets js styles views index.html"
 
 echo "Deploying låfteweb"
@@ -10,4 +10,14 @@ then
     DEPLOY_DIR=$INPUT_DEPLOY_DIR
 fi
 echo "Deploying to $DEPLOY_DIR"
-cp -r -v $DEPLOY_FILES $DEPLOY_DIR
+echo "WARNING! This will overwrite files in the deploy directory. Are you sure you want to continue? (y/n)"
+read CONFIRMATION
+if [ $CONFIRMATION == "y" ]
+then
+    cp -r -v $DEPLOY_FILES $DEPLOY_DIR
+elif [ $CONFIRMATION == "Y" ]
+then
+    cp -r -v $DEPLOY_FILES $DEPLOY_DIR
+else
+    echo "Aborting"
+fi
